move more map-specific recoil selectors into features/map/state
